+++ cmd/migrations/postgres/3_add_user_orders_table.sql
@@ -1,7 +1,7 @@
 CREATE table if not exists "user_orders"
 (
     "user_id"  int  NOT null,
-    "order_id" int  NOT null,
+    "order_id" bigint  NOT null,
     CONSTRAINT fk_user_id FOREIGN KEY ("user_id") REFERENCES users ("user_id"),
     CONSTRAINT fk_order_id FOREIGN KEY ("order_id") REFERENCES orders ("order_id")
 );
