+++ cmd/migrations/postgres/5_add_withdraws_table.sql
@@ -1,7 +1,7 @@
 CREATE table if not exists "withdraws"
 (
     "withdraw_id"  SERIAL PRIMARY KEY NOT null,
-    "order_id"     int                NOT null,
+    "order_id"     bigint                NOT null,
     "sum"          int                NOT null,
     "processed_at" timestamp
 );
